fix spm module directory
