+++ Mutaro.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -44,6 +44,15 @@
         "branch" : "main",
         "revision" : "284a41800b7c5565512ec6ae21ee818aac1f84ac"
       }
+    },
+    {
+      "identity" : "swiftgenplugin",
+      "kind" : "remoteSourceControl",
+      "location" : "https://github.com/SwiftGen/SwiftGenPlugin",
+      "state" : {
+        "revision" : "879b85a470cacd70c19e22eb7e11a3aed66f4068",
+        "version" : "6.6.2"
+      }
     }
   ],
   "version" : 2
