+++ MutaroApp/MutaroApp/AppDelegate.swift
@@ -5,7 +5,7 @@
 //  Created by minguk-kim on 2022/12/29.
 //
 
-import FirebaseSetupModule
+import FirebaseSetup
 import UIKit
 
 @main
