--- MutaroModule/Sources/BuildModule/BuildModule.swift
@@ -1,12 +1,0 @@
-//
-//  BuildModule.swift
-//
-//
-//  Created by minguk-kim on 2022/12/29.
-//
-
-import Foundation
-
-struct BuildModule {
-
-}
