+++ MutaroModule/Sources/Core/Extensions/Applicable.swift
