+++ MutaroModule/Sources/Features/Views/AppIntroductionView/ViewModel/AppIntroductionViewModel.swift
@@ -1,0 +1,15 @@
+//
+//  AppIntroductionViewModel.swift
+//
+//
+//  Created by minguk-kim on 2023/01/01.
+//
+
+import Core
+import UIKit
+
+final class AppIntroductionViewModel: NSObject {
+    func onTapAgree() {
+        UserDefaults.standard.set(true, forKey: UserDefaultsKey.notFirstAppLaunching.rawValue)
+    }
+}
