+++ MutaroModule/Sources/ImageLoader/Utilities/ImageContentPathProvider.swift
