--- MutaroModule/Sources/MutaroClientModule/MutaroRepository/MutaroClient.swift
@@ -1,21 +1,0 @@
-//
-//  MutaroClient.swift
-//
-//
-//  Created by minguk-kim on 2023/01/04.
-//
-
-import CommonAppModule
-import FirebaseFirestore
-import Foundation
-
-public final class MutaroClient {
-    public static let shared = MutaroClient()
-
-    public let firestore = Firestore.firestore()
-    public let networkStatusManager: NetworkStatusManagerProtocol
-
-    private init(networkStatusManager: NetworkStatusManagerProtocol = NetworkStatusManager()) {
-        self.networkStatusManager = networkStatusManager
-    }
-}
