+++ MutaroModule/Sources/Repositories/MutaroRepository/MutaroDetails/MutaroDetailResource.swift
